🛠 Refactored tests
