--- Makefile
@@ -1,7 +1,0 @@
-
-test:
-	go test -cover
-
-cover:
-	go test -cover -coverprofile=c.out
-	go tool cover -html=c.out -o coverage.html
