+++ middlewares/color/color.go
